- fix PagerView shadow issue with add white UIView
