+++ Real/Real/Controller/UserProfileViewController.swift
@@ -12,32 +12,31 @@ import FSPagerView
 
 class UserProfileViewController: BaseViewController {
 
-    @IBOutlet weak var pagerView: FSPagerView! {
-        
-        didSet {
-            
-            self.pagerView.register(FSPagerViewCell.self, forCellWithReuseIdentifier: "cell")
-            
-            pagerView.delegate = self
-            
-            pagerView.dataSource = self
-            
-            pagerView.transformer = FSPagerViewTransformer(type: .overlap)
-            
-            pagerView.itemSize = CGSize(
-                width: pagerView.frame.width * 0.75,
-                height: pagerView.frame.size.height * 0.9
-            )
-        }
-    }
-    @IBOutlet weak var buttonView: UIStackView! {
-        
-        didSet {
-            buttonView.setupBorder(width: 0.8, color: .lightGray)
-        }
-    }
+//    @IBOutlet weak var pagerView: FSPagerView! {
+//        
+//        didSet {
+//            
+//            self.pagerView.register(FSPagerViewCell.self, forCellWithReuseIdentifier: "cell")
+//            
+//            pagerView.delegate = self
+//            
+//            pagerView.dataSource = self
+//            
+//            pagerView.transformer = FSPagerViewTransformer(type: .overlap)
+//            
+//            pagerView.itemSize = CGSize(
+//                width: pagerView.frame.width * 0.75,
+//                height: pagerView.frame.size.height * 0.9
+//            )
+//        }
+//    }
     
     let authManager = FirebaseAuthManager()
+    
+    override var isHideTabBar: Bool {
+        
+        return true
+    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -53,25 +52,25 @@ class UserProfileViewController: BaseViewController {
 //    }
 }
 
-extension UserProfileViewController: FSPagerViewDelegate, FSPagerViewDataSource {
-    
-    func numberOfItems(in pagerView: FSPagerView) -> Int {
-        return 5
-    }
-    
-    func pagerView(_ pagerView: FSPagerView, cellForItemAt index: Int) -> FSPagerViewCell {
-        
-        let cell = pagerView.dequeueReusableCell(withReuseIdentifier: "cell", at: index)
-        
-        cell.setupShadow()
-                
-        cell.backgroundColor = .white
-        
-        return cell
-    }
-    
-    func pagerView(_ pagerView: FSPagerView, didSelectItemAt index: Int) {
-        
-        print("selected index: \(index)")
-    }
-}
+//extension UserProfileViewController: FSPagerViewDelegate, FSPagerViewDataSource {
+//
+//    func numberOfItems(in pagerView: FSPagerView) -> Int {
+//        return 5
+//    }
+//
+//    func pagerView(_ pagerView: FSPagerView, cellForItemAt index: Int) -> FSPagerViewCell {
+//
+//        let cell = pagerView.dequeueReusableCell(withReuseIdentifier: "cell", at: index)
+//
+//        cell.setupShadow()
+//
+//        cell.backgroundColor = .white
+//
+//        return cell
+//    }
+//
+//    func pagerView(_ pagerView: FSPagerView, didSelectItemAt index: Int) {
+//
+//        print("selected index: \(index)")
+//    }
+//}
